Add Empty property and IsEmpty check to User class
Added an Empty static property to the User class in the Authentication library, which is used to represent an absence of a valid user instance. Also, implemented the IsEmpty property to check if the User object is empty or not, determined by whether the Id property is null or an empty string.