Enhance user authentication and add 'GetUsers' method
An enhancement was made to the user authentication process to also check by email, not just by username. A user is authenticated if the entered password matches the encrypted data in the database and the obtained user details is not empty. Additionally, a 'GetUsers' method was added to retrieve all users from the database. In the 'GenerateToken' method, an exception now gets thrown if the provided user is empty.